Fixed warnings in AHI code (Part 1). (I also have a patch to pass the same warning flags to AHI as to other code, but I need to fix the rest of the warnings first.)
git-svn-id: https://svn.aros.org/svn/aros/trunk@52229 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/devs/AHI/AHI/gui_mui.c
@@ -327,7 +327,7 @@ static void GUINewMode(void)
   if( buffer != NULL )
   {
     sprintf( buffer,"0x%08lx\n%s\n%s\n%s\nDevs:AHI/%s.audio\n%s\n%s",
-      getAudioMode(),
+      (unsigned long)getAudioMode(),
       arg1,
       arg2,
       arg3,
